feat: :sparkles: Added Cinema Manager Dashboard + Split up views for redability
