+++ Project/UWEFlix/views/__init__.py
@@ -1,0 +1,5 @@
+from .login import *
+from .booking import *
+from .account import *
+from .account_manager import *
+from .cinema_manager import *
